+++ marker_api/src/sem/ty.rs
@@ -5,6 +5,7 @@ mod ptr_ty;
 mod sequence_ty;
 mod trait_ty;
 mod user_ty;
+
 pub use fn_ty::*;
 pub use other_ty::*;
 pub use prim_ty::*;
@@ -12,6 +13,9 @@ pub use ptr_ty::*;
 pub use sequence_ty::*;
 pub use trait_ty::*;
 pub use user_ty::*;
+
+use crate::common::DriverTyId;
+use std::{fmt::Debug, marker::PhantomData};
 
 /// The semantic representation of a type.
 #[repr(C)]
@@ -43,10 +47,10 @@ pub enum TyKind<'ast> {
     // ================================
     /// A [function item type](https://doc.rust-lang.org/reference/types/function-item.html)
     /// identifying a specific function and potentualy additional generics.
-    FnTy(&'ast FnTy<'ast>),
+    Fn(&'ast FnTy<'ast>),
     /// The semantic representation of a
     /// [closure type](https://doc.rust-lang.org/reference/types/closure.html).
-    ClosureTy(&'ast ClosureTy<'ast>),
+    Closure(&'ast ClosureTy<'ast>),
     // ================================
     // Pointer types
     // ================================
@@ -110,3 +114,43 @@ impl<'ast> TyKind<'ast> {
         ty
     }
 }
+
+#[repr(C)]
+#[cfg_attr(feature = "driver-api", visibility::make(pub))]
+#[cfg_attr(feature = "driver-api", derive(typed_builder::TypedBuilder))]
+pub(crate) struct CommonTyData<'ast> {
+    #[cfg_attr(feature = "driver-api", builder(default))]
+    _lifetime: PhantomData<&'ast ()>,
+    driver_id: DriverTyId,
+}
+
+impl<'ast> Debug for CommonTyData<'ast> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("CommonTyData {...}").finish()
+    }
+}
+
+#[cfg(feature = "driver-api")]
+impl<'ast> CommonTyData<'ast> {
+    pub fn driver_id(&self) -> DriverTyId {
+        self.driver_id
+    }
+}
+
+macro_rules! impl_ty_data {
+    ($self_ty:ty, $enum_name:ident) => {
+        #[cfg(feature = "driver_api")]
+        impl<'ast> $self_ty {
+            pub fn data(&self) -> &$crate::sem::ty::CommonTyData<'ast> {
+                &self.data
+            }
+        }
+
+        impl<'ast> From<&'ast $self_ty> for $crate::sem::ty::TyKind<'ast> {
+            fn from(from: &'ast $self_ty) -> Self {
+                $crate::sem::ty::TyKind::$enum_name(from)
+            }
+        }
+    };
+}
+use impl_ty_data;