+++ marker_api/src/sem/ty/other_ty.rs
@@ -1,16 +1,12 @@
-use std::marker::PhantomData;
+use super::CommonTyData;
 
 /// The placeholder type, signalling that the semantic type is still unstable
 /// and therefor not represented as part of the API.
 #[repr(C)]
 #[derive(Debug)]
+#[cfg_attr(feature = "driver-api", derive(typed_builder::TypedBuilder))]
 pub struct UnstableTy<'ast> {
-    _lt: PhantomData<&'ast ()>,
+    data: CommonTyData<'ast>,
 }
 
-#[cfg(feature = "driver-api")]
-impl<'ast> UnstableTy<'ast> {
-    pub fn new() -> Self {
-        Self { _lt: PhantomData }
-    }
-}
+super::impl_ty_data!(UnstableTy<'ast>, Unstable);