+++ marker_api/src/sem/ty/sequence_ty.rs
@@ -1,11 +1,14 @@
 use crate::{ffi::FfiSlice, sem::ConstValue};
 
-use super::TyKind;
+use super::{CommonTyData, TyKind};
 
 /// The semantic representation of a tuple type like [`()`](prim@tuple) or [`(T, U)`](prim@tuple)
 #[repr(C)]
 #[derive(Debug)]
+#[cfg_attr(feature = "driver-api", derive(typed_builder::TypedBuilder))]
 pub struct TupleTy<'ast> {
+    data: CommonTyData<'ast>,
+    #[cfg_attr(feature = "driver-api", builder(setter(into)))]
     types: FfiSlice<'ast, TyKind<'ast>>,
 }
 
@@ -15,12 +18,7 @@ impl<'ast> TupleTy<'ast> {
     }
 }
 
-#[cfg(feature = "driver-api")]
-impl<'ast> TupleTy<'ast> {
-    pub fn new(types: &'ast [TyKind<'ast>]) -> Self {
-        Self { types: types.into() }
-    }
-}
+super::impl_ty_data!(TupleTy<'ast>, Tuple);
 
 impl<'ast> std::fmt::Display for TupleTy<'ast> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
@@ -36,7 +34,9 @@ impl<'ast> std::fmt::Display for TupleTy<'ast> {
 
 /// The semantic representation of a variable length slice like [`[T]`](prim@slice)
 #[repr(C)]
+#[cfg_attr(feature = "driver-api", derive(typed_builder::TypedBuilder))]
 pub struct SliceTy<'ast> {
+    data: CommonTyData<'ast>,
     inner_ty: TyKind<'ast>,
 }
 
@@ -46,12 +46,7 @@ impl<'ast> SliceTy<'ast> {
     }
 }
 
-#[cfg(feature = "driver-api")]
-impl<'ast> SliceTy<'ast> {
-    pub fn new(inner_ty: TyKind<'ast>) -> Self {
-        Self { inner_ty }
-    }
-}
+super::impl_ty_data!(SliceTy<'ast>, Slice);
 
 impl<'ast> std::fmt::Debug for SliceTy<'ast> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
@@ -62,7 +57,9 @@ impl<'ast> std::fmt::Debug for SliceTy<'ast> {
 /// The semantic representation of an array with a known size like: [`[T; N]`](prim@array)
 #[repr(C)]
 #[derive(Debug)]
+#[cfg_attr(feature = "driver-api", derive(typed_builder::TypedBuilder))]
 pub struct ArrayTy<'ast> {
+    data: CommonTyData<'ast>,
     inner_ty: TyKind<'ast>,
     len: ConstValue<'ast>,
 }
@@ -77,12 +74,7 @@ impl<'ast> ArrayTy<'ast> {
     }
 }
 
-#[cfg(feature = "driver-api")]
-impl<'ast> ArrayTy<'ast> {
-    pub fn new(inner_ty: TyKind<'ast>, len: ConstValue<'ast>) -> Self {
-        Self { inner_ty, len }
-    }
-}
+super::impl_ty_data!(ArrayTy<'ast>, Array);
 
 impl<'ast> std::fmt::Display for ArrayTy<'ast> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {